+++ example/lib/main.dart
@@ -1,7 +1,41 @@
 import 'package:call_log/call_log.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:workmanager/workmanager.dart';
 
-void main() => runApp(MyApp());
+///TOP-LEVEL FUNCTION PROVIDED FOR WORK MANAGER AS CALLBACK
+void callbackDispatcher() {
+  Workmanager.executeTask((task, inputData) async {
+    print('Background Services are Working!');
+    try {
+      Iterable<CallLogEntry> cLog = await CallLog.get();
+      print('Queried call log entries');
+      for (CallLogEntry entry in cLog) {
+        print('-------------------------------------');
+        print('F. NUMBER  : ${entry.formattedNumber}');
+        print('C.M. NUMBER: ${entry.cachedMatchedNumber}');
+        print('NUMBER     : ${entry.number}');
+        print('NAME       : ${entry.name}');
+        print('TYPE       : ${entry.callType}');
+        print('DATE       : ${DateTime.fromMillisecondsSinceEpoch(entry.timestamp)}');
+        print('DURATION   : ${entry.duration}');
+        print('ACCOUNT ID : ${entry.phoneAccountId}');
+        print('SIM NAME   : ${entry.simDisplayName}');
+        print('-------------------------------------');
+      }
+      return true;
+    } on PlatformException catch (e, s) {
+      print(e);
+      print(s);
+      return true;
+    }
+  });
+}
+
+void main() {
+  runApp(MyApp());
+  Workmanager.initialize(callbackDispatcher, isInDebugMode: true);
+}
 
 /// example widget for call log plugin
 class MyApp extends StatefulWidget {
@@ -55,6 +89,21 @@ class _MyAppState extends State<MyApp> {
                       });
                     },
                     child: const Text('Get all'),
+                  ),
+                ),
+              ),
+              Center(
+                child: Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: RaisedButton(
+                    onPressed: () {
+                      Workmanager.registerOneOffTask(
+                        DateTime.now().millisecondsSinceEpoch.toString(),
+                        'simpleTask',
+                        existingWorkPolicy: ExistingWorkPolicy.replace,
+                      );
+                    },
+                    child: const Text('Get all in background'),
                   ),
                 ),
               ),